SINGA-235 - Unify the engines for cudnn and singa layers
Fixed a bug in alexnet.cc caused by forgeting to udpate the layer
construction code.
Updated some code to avoid the warnings from the compilation.

+++ test/singa/test_image_transformer.cc
@@ -61,7 +61,7 @@ TEST(ImageTransformer, Apply3D) {
   srand(time(NULL));
   for (size_t i = 0; i < n; i++) x[i] = (float)(rand() % 256);
   in.CopyDataFromHostPtr<float>(x, n);
-  size_t resize_height = 4, resize_width = 6;
+  int resize_height = 4, resize_width = 6;
 
   singa::ImageTransformer img_transformer;
   singa::TransformerConf conf;
@@ -90,8 +90,8 @@ TEST(ImageTransformer, Apply3D) {
   EXPECT_EQ(resize_width, resized.size().width);
   size_t new_size = resize_height * resize_width * channel;
   float* xt = new float[new_size];
-  for (size_t i = 0; i < resize_height; i++)
-    for (size_t j = 0; j < resize_width; j++)
+  for (int i = 0; i < resize_height; i++)
+    for (int j = 0; j < resize_width; j++)
       for (size_t k = 0; k < channel; k++)
         xt[i * resize_width * channel + j * channel + k] = resized.at<cv::Vec3f>(i, j)[k];
   for (size_t c = 0; c < 3; c++)
@@ -128,7 +128,7 @@ TEST(ImageTransformer, Apply2D) {
   srand(time(NULL));
   for (size_t i = 0; i < n; i++) x[i] = (float)(rand() % 256);
   in.CopyDataFromHostPtr<float>(x, n);
-  size_t resize_height = 4, resize_width = 6;
+  int resize_height = 4, resize_width = 6;
 
   singa::ImageTransformer img_transformer;
   singa::TransformerConf conf;
@@ -156,8 +156,8 @@ TEST(ImageTransformer, Apply2D) {
   EXPECT_EQ(resize_width, resized.size().width);
   size_t new_size = resize_height * resize_width;
   float* xt = new float[new_size];
-  for (size_t i = 0; i < resize_height; i++)
-    for (size_t j = 0; j < resize_width; j++)
+  for (int i = 0; i < resize_height; i++)
+    for (int j = 0; j < resize_width; j++)
         xt[i * resize_width + j] = resized.at<cv::Vec<float, 1>>(i, j)[0];
 
   for (size_t h = 0; h < 2; h++)
@@ -187,7 +187,7 @@ TEST(ImageTransformer, Resize) {
   srand(time(NULL));
   for (size_t i = 0; i < n; i++) x[i] = (float)(rand() % 256);
   in.CopyDataFromHostPtr<float>(x, n);
-  size_t resize_height = 4, resize_width = 5;
+  int resize_height = 4, resize_width = 5;
   singa::Tensor out = singa::resize(in, resize_height, resize_width, "HWC");
   const float* y = out.data<float>();
 
@@ -203,8 +203,8 @@ TEST(ImageTransformer, Resize) {
   EXPECT_EQ(resize_width, resized.size().width);
   size_t new_size = resize_height * resize_width * channel;
   float* xt = new float[new_size];
-  for (size_t i = 0; i < resize_height; i++)
-    for (size_t j = 0; j < resize_width; j++)
+  for (int i = 0; i < resize_height; i++)
+    for (int j = 0; j < resize_width; j++)
       for (size_t k = 0; k < channel; k++)
         xt[i * resize_width * channel + j * channel + k] = resized.at<cv::Vec3f>(i, j)[k];
 
@@ -232,7 +232,7 @@ TEST(ImageTransformer, Crop) {
     for (size_t w = 0; w < crop_width; w++)
       for (size_t c = 0; c < channel; c++) {
         size_t out_idx = c * crop_height * crop_width + h * crop_width + w;
-        size_t in_idx = c * height * width + (h + crop_h_offset) 
+        size_t in_idx = c * height * width + (h + crop_h_offset)
                  * width + w + crop_w_offset;
         EXPECT_EQ(x[in_idx], y[out_idx]);
       }
